Update snapshots for K1 to not contain KTS modules references
With IntelliJ 2025.2 the Gradle KTS are added as project modules,
happening this behaviour from KotlinDslScriptSyncContributor. However,
this is only applicable for K2 since for K1 the old behaviour is
kept given that this mode is already deprecated and will soon be
removed. For this reason, now test projects using KTS will require
different snapshot files one for K1 and another for K2.

Bug: 450612068
Test: This
Change-Id: I6df76914c3e10779d2ea3690b70dff8619f3a062