[lint] Don't disregard superclass intentions
Even if a plugin registers a dependency update provider, it probably
shouldn't override the base intentions (if any).

As it happens there currently aren't any intentions from the
superclass, but there might be and it will probably be less confusing
eventually to make sure that they do not get overridden based on the
presence or absence of a plugin.

Bug: 454720142
Test: existing (no direct change in behaviour)
Change-Id: Ifd5faac51a229e76a6bd76eacb77d9c0aef901c2