Refactor Preview Details with common img toolbar
Replaces the multiple, individual image toolbars in the single-preview "All" view with a single, common toolbar.
This provides a cleaner UI and a more intuitive UX, as actions like zoom and grid now apply to all visible images simultaneously.
This change refactors the panel to use a new ImageWithToolbarPanel component, modeled after the one in ScreenshotResultView, to encapsulate image state and controls.

Bug: 449797945
Test: N/A. Manual.
Change-Id: Ie441f0ede4ed05fa3d981578c0790ebb377d0b6c