[AUA] Distinguish between incompatibility and unsupportedness
Now that we have the "forced upgrade" compatibility state -- which
effectively consists of a successful sync with a callback -- we need
to distinguish between a formal incompatibility (which would most
likely lead to an unsuccessful sync) and an AGP version that is
compatible but considered unsupported.

Where this matters is not visible in this change: we preserve the
intended meaning of `versionsAreIncompatible`, used by the project
startup activity to determine whether it can reopen a project from
cached models rather than triggering sync; the new
`versionsAreUnsupported` function with the existing implementation is
used in the AGP Upgrade Assistant to filter unsupported versions from
the list of suggested versions to upgrade to.

This behaviour change was picked up while trying to adapt Ndk synced
tests to work with OBSOLETE-state AGP versions.

Bug: none filed
Test: minor adjustments of existing tests
Change-Id: I98706abf3bc9b0b7214f9d867a2bd8f06b75cc01