[AUA] Disable the forced upgrade dialog in snapshot tests
Rather than mocking, allow subclassing
RefactoringProcessorInstantiator; use this to provide a utility
function which replaces the service with a fake, and also extend
OpenProjectOptions in the test framework to allow a boolean option for
disabling it.

Adapt existing `open` calls to the new flag, and disable the dialog in
SyncedProjectTestBase.

Bug: none filed
Test: existing (effectively test-only change)
Change-Id: I5f28a9e28a87b3c0fe5441c69870740602c79698