Fix flaky test: StateInspectionPanelIntegrationTest
This test revealed a problem with the model:

When a new request is emitted via stateReadRequested the cache
could react and send a (cached) stateReads result before the model
has a chance to handle the new stateReadRequested.

If the model handles the stateRead result before the stateReadRequest
the model may overwrite the result with a waiting state. Detect this
in the model.

Fixes: 458142095
Test: StateInspectionPanelIntegrationTest
Change-Id: I387ec6e03d63e8980eec5eb58e254ca0f3e9d67f