Fix SS attributes failed state text & scroll
The view's content and the scrollbar have been wrapped in a Row to position the scrollbar correctly on the right.
When a screenshot test failed without a calculated diff percentage, the "Match" field would display the text "FAILED".
This has been changed to show a more intuitive "0.00%" to indicate a complete mismatch.

Bug: 457924729
Test: Existing unit tests in ScreenshotAttributesViewTest.kt pass &
Manual verifiction.

Change-Id: I0a91e999d852202f52f33b099a950bb16b72b947