Support injecting MavenRepo without init-script
With new project templates and the AGP 8.3, Android Studio generates a
`settings.gradle.kts` which disregards the legacy `buildscript` usage the in
the init-script and fails the sync. While our test projects don't use the
`FAIL_ON_PROJECT_REPOS` mode yet, `CreateProjectTest` is failing
with the version upgrade repro – ag/36872630.
We have added a method to skip the init-script usage, while keeping the
default behavior same.

Bug: 459349293
Test: Manually with CreateProjectTest and ag/36872630
Change-Id: I5921ddb31f9d99cf31a876191f92ce65d06ee631