Add retry mechanism to AndroidGradleProjectOpenProcessor.canOpenProject
With IntelliJ 2025.3 when importing project in Studio the
AndroidGradleProjectOpenProcessor fails to detect it as Gradle
project despite having defined Gradle build files, which doesn't
longer trigger the GradleProjectImporter to create and configure it.

This is caused because the VirtualFile.refresh doesn't properly
update the file information from the physical file were some
children are missing.

Bug: 459390354
Test: Existing
Change-Id: Iff019fcc3196e1538a4b9fdd28ecc2a58302f848