Add Predictive Pack navigation controls in Interactive mode
This change introduces a UI panel for controlling back navigation when
using Interactive Preview Mode.

The new UI is implemented in NavigationControlsPanel and adds a panel
with the following controls:
* Back event via a button (overflow menu on the top right).
* Predictive Back progress simulation.
* Predictive Back edges simulation.

To support this, InteractiveNavigationHandler is refactored into
InteractivePreviewNavigationController. This controller manages the
visibility of the new navigation panel.

It also uses reflection to interface with Android's back navigation
system where it is possible to call the Navigation Event APIs: start,
progress, completion, and cancellation.

Bug: 462365661
Test: BottomPanelNavigationControllerTest
Test: BottomNavigationPanelUiTest
Test: existing updated
Change-Id: Id2f9a12eeae1e3f7f69a515f98a289833a9c425f