Do not dispatch output events when not necessary
When gradle execution is requested via AndroidGradleTaskManager from
internals of the IJ platform (which is the case for running unit tests)
it comes with the listener that internally notifies events dispatcher.
As we are manually notifying the dispatcher here, it gets notified
twice for each event which results in duplicated lines.

startBuildEventPosted is a good indication of if this is needed or not,
build output infrastructure being properly set up before us here also
means dispatcher is being handled correctly already.

Fixes: 301232493
Test: Existing integration tests should catch any regressions, manually
verified that duplicated output issue is fixed with this.
Change-Id: I1941261a49ab594230f02ca3ee3f8b248df37b5d