Update which old AGP version we test in ProviderOldAgpIntegrationTest
As we now moved to minimum version being AGP 4.2, we remove all the
assertions for old AGPs (AGP 3.5 was just there anywaybut not tested).

As we now will also be forcing upgrade for AGP <4.2, update the test to
run against AGP >=4.2 only.

For clarification: while AGP 3.5 is completely unsupported by AS and we
removed tests for it, AGP 4.0 and 4.1 are in the "forced upgrade" phase,
so this means that we still need to keep the tests that run Sync to
verify its behavior, but it should be safe to get rid of the tests that
verify the deployment results, as deployment is unsupported for these
versions anyway.

Bug: 460371674
Test: this
Change-Id: I05e244bd28deeaa05808eb22c266fa49d1809ded