Use gradle tooling events for receiving tests
IJ moved to using Gradle TAPI events for tests some time ago, see
GradleTaskManager.setupBuiltInTestEvents. In AS we have separate
implementation of GradleTasksExecutor which did not adopt same change,
do this now.

Add BuildOutputRunningFromRunConfigurationTest to test result observed
on build output. It is important that in this test we run gradle test
using run configuration, this ensures full process of configuring
gradle invocation as it happens for the user and allows to catch
problems not visible on other tests. This test verifies output state
for b/441472103 and b/301232493 fixed earlier.

Fixes: 441472103
Bug: 301232493
Test: BuildOutputRunningFromRunConfigurationTest
Change-Id: Ie7053905e779d34c415851774a748ca876d3bcb5