Hide an action in Gradle projects
The AssembleIdeaModuleAction almost always hides itself in any case;
it would only be active when the project system is a
GradleProjectSystem, but without a requirement for Android models (so,
a pure Java Gradle project?).  In those circumstances, the
AssembleGradleModuleAction is active, so there seems to be no benefit
to this action.

The 10-year-old comment about not hiding this action because of an NPE
appears to be obsolete.

Bug: 452574547
Test: existing
Change-Id: I880d5dd14064a670b8db71a86aafbc869ca829e7