Fix stack overflow in `AndroidJdiHelperClassLoader`
With the new platform update Intellij Java debugger started using the method invocation helper class more frequently during evaluation.
When `AndroidJdiHelperClassLoader` tried to load the mentioned method invocation helper, it used the evaluation machinery to do so.
At the same time, the evaluation mechanism tried to use this helper  class for method invocation, which triggered its loading
with `AndroidJdiHelperClassLoader`. This caused stack overflow. This commit fixes the issue by explicitly telling the debugger
to not use method helper when evaluating inside of `AndroidJdiHelperClassLoader`.

Test: N/A
Bug: 452296419
Change-Id: I3ef284e209c541dd7232679e96038ffdbb996e0b