RAG: Enable new RAG Lucene indexer
The current RAG indexer used for code_search tool of the agent relies on
the IntelliJ file-based indexer, but it generates the Lucene index for
BM25 as a side effect, which means it does not update the index when the
file is deleted while the IDE is not running. It has other cases of
unsynced index issues like the kept VirtualFile ID is out of the range.
The new RAG Lucene indexer will fix this issue. We already tested it
by setting the flag as true more than a week. We want to enable it for
Canary.

Bug: 432317145
Test: N/A
Change-Id: I940b7b4ee2fe3bbc3344f245bea0682db6ee4278