Fix TestSuiteUtilsTest with IntelliJ 2025.3
The TestSuiteUtilsTest defines a test suite where 'myTestSuite'
and 'myTestSuiteWithMultipleTargets' both are pointing to the same
content roots directory 'src/myTestSuite' and obtaining a linked
testSuiteModule for added file under that directory using
ModuleUtilCore.findModuleForFile().

With IntelliJ 2025.3 platform addressed an issue in the WSM, which
now given a file added on that content root, it returns both modules
'myTestSuite' and 'myTestSuiteWithMultipleTargets', being this the
expected behaviour.

All of this makes the ModuleUtilCore.findModuleForFile() ambiguous
since it picks the first one it finds making tests to fail. The
ModuleUtilCore.findModulesForFile() should be used instead filtering
by the expected test suite module.

Since this scenario is completely out of scope for this test, a
separate one will be created for its validation ensuring platform
expected behaviour in such case.

Bug: 459426422
Test: TestSuiteUtilsTest
Change-Id: I409abd37edc9f626b9d134894a2426f62af30f93