Remove aswb.auto.convert.legacy.sync.to.query.sync experiment
This experiment has been rolled out to 100 since cl/763855325.
The usages of the flag have been changed in the code to be assumed
to always be true and where that would make the result always
true (e.g. || FLAG) or always false (e.g. && !FLAG) the code has
been replaced with the resulting boolean constant.

Bug: 447150739
Test: N/A
Change-Id: I5bb146e2fa98610220ee6dbb9c7dfc36038e6b01