Query sync knows which proto runtime is used
with each file. For .java/.kt files it knows which proto runtime it
depends on and for .proto files it knows which kinds of targets use it
in the current project view.

Note:
Currently it reflects the build dependency graph similarly to what
on-demand dependency analysis did and therefore it is still possible for
a .java/.kt file to report both full and lite runtimes. However, these
queries will start returning the current sync state instead and since
every /java/kt file can be synced in the context of one target and
configuration only they will start returning only one runtime.

Remove no longer needed on demand graph traversal in the protobuf
support integration code and instead consult the QuerySyncProject.

Bug: n/a
Test: ProtoModeIntegrationTest
Test: SyncedInBazelProjectTest.simpleProto
Change-Id: I5b4fe5827059144b6f4b85abc167e3e080a6b903