[KMT-1781][android][preview] fix multipreviews in KMP context
Most of the parts of previews resolving are implemented with UAST, which is supposed to support only Kotlin JVM, not KMP. Because of that the annotations of e.g. multipreview annotations were not collectd properly, and their arguments were not being extracted. This change fixes that stuff at leaves of the operations performed, where the internal model stays represented with UAST abstractions.

In the ideal world, the whole machinery should be migrated to use the proper Kotlin-centric abstractions, instead of UAST, to support KMP natively.

GitOrigin-RevId: 7acf4fbe68ad213c92e99a0c7a78977de857a405
(cherry picked from commit 5efa0ab01af948c66c02f3280f593ecf88d6ee59)
Change-Id: Ie8662f21b6b2bc4d54d6c2551e1dbed6396dd8bd