Organise remaining classes in 'deprecated' directory
The form classes define the UI of the wizard and are used by the newer
wizard model, so they've been moved into a new 'form' package.

The `WelcomeScreenWindowListener` + `CancelableWelcomeWizard` classes
are also used by the new wizard to override the default IntelliJ
behaviour when the welcome frame is closed, so have been moved into the
parent package.

The statics in `InstallComponentsPath` have been moved into the classes
in which they are referenced.

Bug: 463955218
Test: n/a
Change-Id: I132530a2259e98013661e10211e178c7d6e1d8f8