Adapt skip event ordering on BuildListenerTest with IntelliJ 2025.3
With IntelliJ 2025.3 platform introduced ProjectSettingsEntity for
project SDK and JavaProjectSettingsEntity for java LanguageLevel
and compiler outputs, which was done as part of IJPL-15953 ticket
on bcfbe7570460 commit.

However, re-opening an imported project is now causing a re-scan when
comparing the old and new WorkspaceEntity, which results on
ModuleRootListener.changeRoots event to be triggered before the
GradleSyncListenerWithRoot.syncSkipped. This is because
ProjectSettingsEntity always triggers a re-indexed based on
CustomEntitiesCausingReindexTracker.shouldCauseRescan implementation:

  boolean shouldRescan(WorkspaceEntity old, WorkspaceEntity new) {
    if (anyEntity instanceof ProjectSettingsEntity) {
     return true; // project sdk is always indexed
    }
  }

Despite this not being the previous behaviour, isn't expected to cause
any issue based on other tests we have. Moreover, while debugging this,
found that re-opening a project in BuildListenerTest test triggers a
full JPS project load, while in production a partial update is done
Consequently, production does not trigger a rootsChanged event when
reopening an already imported project if no modification happened.

Bug: 459420307
Test: BuildListenerTest
Change-Id: I32d3aa0d5a4d6e131830319b78d13f2f486ff79d