Work around missing information from the KotlinModelBuilder
We do not pass the needed srcSet information in the kotlincompilation
data to the
KotlinGradleModelBuilder in order to setup visibility onto other sourceSets.
Because of this, and in order to be able to add support for having
visibility onto "friends" sourceSets, we will be setting this
information at the level of the kotlin facet configuration (in the data
service).

Bug: 357930912
Test: reflected on the snapshot tests
Change-Id: I715fec777475c1835f0c4352f394da0c1c8c8c5e