Remove '_phased' suffix from baseline snapshots files
The '_phased', '_K2_phased' and '_K2' suffixes were created to be
added only in those cases when exist difference between baseline
and K2 and/or phase sync options are enabled.

Now, if there's no difference ideally those should be named without
any suffix being this '_V2.txt'. The reason of this is because,
when disabling that option then the baseline will no longer be used,
requiring to have unnecessarily different file.

Bug: 460369658
Test: Existing
Change-Id: I9ed329e67cb9edc23449e7ec40a71234867100f1