Fix PreviewModeManager when background changes
When the background changes, all the preview element definitions are
considered to have changed. This causes focus mode to consider all new
and pick the first one available.
With this fix, we use the heuristic of preferring the preview element
definition that retains the same annotation definition PSI element.

This heuristic will apply after all existing ones so it will impact only
the cases where we could not make a reasonable selection before.

Test: PreviewModeManagerTest
Fixes: 464452310
Change-Id: I758d03513f9cf750bdefb95582251540bd373553