Tweak File Name Handling
* Add ".logcat" extension if not specified when saving.
* Allow importing files with arbitrary extensions.

This is because of changes in the behavior of the File Choosers on
Windows machines.

1. On Linux, the extension isn't displayed on the edit field and is
added automatically. On Windows, it is shown and if removed by user,
isn't added.

2. On Linux, only files matching the filter are displayed. On Windows,
all files are displayed but if a file not maching the filter is
selected, it results in an error popup.

To make the behavior consistent, I removed the filter on all platforms.

Fixes: 464439823
Test: Updated
Change-Id: Ib0026cd60703e5d7a26df379e6dc89eb86a4496a