Close dialog on build or execution failure
When updating reference images, the dialog would previously hang if the
build or execution failed before any tests were discovered. To address
this, an `ExecutionListener` is added to detect process termination.
If the run configuration for updating reference images fails
(indicated by a non-zero exit code) and no tests have been discovered yet, the `UpdateReferenceImagesDialog` is
now automatically closed. Additionally, the "Run" tool window is
activated, allowing the user to view the build or execution errors.

Bug: 461374271
Test: Verified manually on android studio
Change-Id: I4937d09936d632733918a0b44a401b274e0af106