Add support for checking JBTable text in E2E tests
Swing uses cell rendering to render a JTable, which uses a single
component to "paint" each cell onto the screen. As a result, there are
no components to match against when we want to validate the UI of a
table.

Instead, we can update the logic in `getTextFromComponent` to return a
textual representation of the whole table, which allows us to use the
`addComponentTextContainsMatch` matcher to validate the table contents.

Presubmit-Runs-Per-Test: //tools/vendor/google/ml/aiplugin/android/journeys/integration:JourneysIntegrationTest.*@100
Test: JourneysIntegrationTest
Bug: 465359627
Change-Id: I88069872e5945335b63279552bd1a117028009c2