Fix indexing loop when too many DWFs are reparsed
When there are many DWFs we end up in an indexation loop.

The manifest being merged triggers a reparse of DWFs. When there are 20+
files reparsed, it triggers a re-index which in turn triggers the
manifest from being re-merged, which then triggers a new reparse, etc.

This CL adds a key to identify which DWFs have a schema loaded for them.
We only reparse those files.

Fixes: 464216127
Test: existing + manually verified on a project with 20+ DWFs
Change-Id: I06d3effef32bd212950909b063ffe8c52cb13156