Delay resolving delegate action
to avoid early action instantiations that may crash.

This change introduces a workaround. When replacing actions
conditionally we request the original action at the time of replacement
and if we receive a stubbed action, we reregister it under a different
id so that we can still unstub it later even after we replace the action
associated with the main id.

Bug: 464737975
Test: manually
Change-Id: I46422c8c50f5dcad70cbd2174283e2ceabc61edc