Enable new Journeys wizard in stable
This flag controls whether the new wizard, which has an updated UI and
is no longer tied to the 'NewAndroidComponentAction' framework, is
shown instead of the legacy wizard.

The wizard will still only be displayed when the
`JOURNEYS_WITH_GEMINI_EXECUTION` flag is also enabled (i.e. the flag
controlled by the Studio Labs feature).

Test: NewAndroidComponentActionTest + TemplateResolverTest
Bug: 465359627
Change-Id: I2382a725dd78e2007ad14d889652fbae3104d2aa