[Phased Sync] Use missingDimensionStrategy in variant selection
To resolve the conflicts of dimension mismatch between dependent
projects.
When a project declares a missingDimensionStrategy for a missing
dimension, then this gets used when resolving projects dependencies that
have the dimension defined.
If there is no missingDimensionstrategy defined and there is no conflict
in resolution, then we fallback to the productFlavor we could resolve.

The behavior of this change replicates the exact logic we currently have
in AGP, by relying on the declared dependencies model here.

Bug: 467047467
Test: N/A
Change-Id: I05c7cdb49e0dd81a67e9a0ca4970ad20efd6d362