Compress zip to work around 4 GB limit
This zip only gets used in E2E tests, so it seems OK to compress it.
It will almost always be cached anyway because the files come from a
pinned IntelliJ release.

Note, there are clues we should not be using
@bazel_tools//tools/zip:zipper in the first place (e.g. see comments
in https://github.com/bazelbuild/bazel/issues/23662). In the long term
we could migrate to some other zipper tool or write our own. However,
some of the obvious choices such as @rules_pkg do not preserve UNIX
file permissions.

Bug: 467773803
Test: Bazel build with aiplugin targeting IntelliJ 2025.3
Change-Id: Ie8a2bbf27ad7dfe79e375c895d1094654248865d