Disable the action to build Run Configs for non-android Projects
As per the initial design of this new BuildAction, we had agreed to
build only Android RCs to begin with.
It would be a nice addition to be able to build the Run Configs for JVM
projects too since we already provide other build actions (project,
module, test scopes for example), but we can do this (if we decide to)
in a following change.

Bug: 467659028
Test: Added
Change-Id: Idbc7fdf0de0d38a4a0d9ec20e5acdd2df046b5d4