Do not fail profiler test if app deployment does not complete before profiling starts.
In a small percentage of cases, the GradleApkBuilder and DeployTask
does not pick up the deployment of the test app on the emulator.
There are other infra issues like gradle build not completing
that also contribute to the issue. These issues are not typical
profiler issues but indicate an infra failure.

This change introduces a wait/check for the app install to start,
and if it fails it skips the profiler test instead of failing.

Bug: 457588735
Test: Startup E2E tests pass consistently
Change-Id: I11c7ec4a78520bd00453b7c58b1067d5105d2361